adding start of work for update/check, need to make prettier
