+++ caliper/client/update.py
@@ -1,0 +1,57 @@
+__author__ = "Vanessa Sochat"
+__copyright__ = "Copyright 2020-2021, Vanessa Sochat"
+__license__ = "MPL 2.0"
+
+from caliper.metrics import MetricsUpdater
+from caliper.logger import logger
+from caliper.utils.file import read_yaml
+import os
+
+
+def main(args, extra):
+
+    # The config file must exist if packages are not defined
+    if not args.packages and (not args.config or not os.path.exists(args.config)):
+        logger.exit(
+            "You must provide an existing caliper.yaml config with --config, or --packages."
+        )
+
+    # Ensure that all metrics are valid
+    client = MetricsUpdater(quiet=True)
+    metrics = args.metric.split(",")
+
+    # If asking for all, we will do all regardless of other specifications
+    if "all" in metrics:
+        metrics = list(client.metrics)
+
+    for metric in metrics:
+        if metric not in client.metrics:
+            logger.exit("%s is not a known metric." % metric)
+
+    # prepare top level output directory
+    outdir = args.outdir or os.getcwd()
+
+    # Read packages from the config, if provided
+    packages = args.packages
+    if args.config:
+        config = read_yaml(args.config)
+        packages = [x["name"] for x in config.get("metrics", []) if x.get("name")]
+
+    # Check or Update the package metrics
+    if args.check:
+        missing = client.check_metrics(packages, metrics, outdir, quiet=True)
+        for package, metrics in missing.items():
+            if not metrics:
+                logger.info("[👀️  ] %s|%s is not found." % (package, metric))
+                continue
+            for metric, versions in metrics.items():
+                if not versions:
+                    logger.info("[✔️  ] %s|%s is up to date." % (package, metric))
+                else:
+                    logger.info(
+                        "[✖️  ] %s|%s has %s new versions."
+                        % (package, metric, len(versions))
+                    )
+
+    else:
+        client.update_metrics(packages, metrics, outdir)
