+++ caliper/managers/git.py
@@ -58,7 +58,7 @@ class GitManager:
 
         # Destination folder can default to present working directory
         dest = dest or self.folder or ""
-        self.run_command(["git", "clone", repo, dest])
+        self.run_command(["git", "clone", "--depth", "1", repo, dest])
         return dest
 
     def init(self, dest=None):
